Move naf code file to common constants
